+++ examples/cpi/simple-cpi-6/trident-tests/fuzz_0/test_fuzz.rs
@@ -22,15 +22,6 @@ impl FuzzDataBuilder<FuzzInstruction> for InstructionsSequence {
     middle_sequence!();
     post_sequence!();
 }
-/// `fn fuzz_iteration` runs during every fuzzing iteration.
-/// Modification is not required.
-fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
-    fuzz_data: FuzzData<T, U>,
-    config: &Config,
-    client: &mut impl FuzzClient,
-) {
-    let _ = fuzz_data.run_with_runtime(client, config);
-}
 fn main() {
     let program_callee = ProgramEntrypoint::new(
         pubkey!("HJR1TK8bgrUWzysdpS1pBGBYKF7zi1tU9cS4qj8BW8ZL"),
@@ -44,5 +35,5 @@ fn main() {
     );
     let config = Config::new();
     let mut client = TridentSVM::new_client(&[program_callee, program_caller], &config);
-    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : InstructionsSequence | { fuzz_iteration (fuzz_data , & config , & mut client) ; });
+    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : InstructionsSequence , client : TridentSVM , config : Config |);
 }