+++ CHANGELOG.md
@@ -14,6 +14,9 @@ incremented upon a breaking change and the patch version will be incremented for
 
 - Trident SVM + AFL (see the PR for more details) ([234](https://github.com/Ackee-Blockchain/trident/pull/234))
 
+**Removed**
+- removed fuzz_iteration from test_fuzz.rs ([243](https://github.com/Ackee-Blockchain/trident/pull/243))
+
 ## [0.8.1] - 2024-11-14
 
 **Removed**